+++ lib/main.ts
@@ -1,3 +1,3 @@
 export { BiznicheProvider } from "./provider";
-export { useGetFeedQuery } from "./hooks/useGetPostsQuery";
+export { useGetFeedQuery } from "./hooks/useGetFeedQuery";
 export { useGetApiViewerQuery } from "./hooks/useGetApiViewerQuery";
